+++ Source/ActivityRunner/Viewer3D/Viewer.cs
@@ -686,7 +686,7 @@ namespace Orts.ActivityRunner.Viewer3D
                 Camera.AttachedCar.Train.FormationReversed = false;
                 (Camera as TrackingCamera).SwapCameras();
             }
-            Simulator.Update((float)elapsedTime.ClockSeconds);
+            Simulator.Update(elapsedTime.ClockSeconds);
             if (PlayerLocomotive.Train.BrakingTime == -2) // We just had a wagon with stuck brakes
             {
                 LoadDefectCarSound(PlayerLocomotive.Train.Cars[-(int)PlayerLocomotive.Train.ContinuousBrakingTime], "BrakesStuck.sms");
